+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This is an alpha version! The changes listed here are not final.
 - Add newsletter access level to the post edit page
 - Forms: Introduce Multiple Choice and Single Choice style variations
 - Remove upgrade nudge for premium block
+- Replace photon implementation with the image-cdn package
 - Update Forms pattern modal default view to Grid
 
 ### Improved compatibility